Move upsert helper function into dbUtil file
